refactor: CommunityCommentsPage 컴포넌트 분리
- CommentLoadingDots: 웨이브 점 로딩 애니메이션 컴포넌트
- CommentItem: 날짜 포맷, @멘션 파싱, 댓글 단일 아이템
- CommentSortButton: 정렬 버튼 + 드롭다운 모달 (외부 클릭 닫기 포함)
- CommentInput: 실제 textarea + 등록 버튼 UI로 교체
- CommunityCommentsPage: 396줄 → 125줄로 축소
