+++ src/components/community/CommentItem/index.ts
@@ -1,0 +1,1 @@
+export * from './CommentItem'
