+++ src/components/community/CommentLoadingDots/index.ts
@@ -1,0 +1,1 @@
+export * from './CommentLoadingDots'
